Actualizacion de los scripts

+++ InstalacionNemoLinux.sh
@@ -1,32 +1,32 @@
 #!/bin/bash
 
-VIM="which vim"
-vacio=""
+opcion=
 
-sudo apt purge nautilus gnome-shell-extension-desktop-icons -y
+if [ "$(whoami)" == "root" ]; then
+  sudo apt purge nautilus gnome-shell-extension-desktop-icons -y
+  sudo apt install nemo -y
+  xdg-mime default nemo.desktop inode/directory application/x-gnome-saved-search
+  settings set org.gnome.desktop.background show-desktop-icons false
+  gsettings set org.nemo.desktop show-desktop-icons true
+  echo "[Desktop Entry]" > /etc/xdg/autostart/nemo-autostart.desktop
+  echo "Type=Application" >> /etc/xdg/autostart/nemo-autostart.desktop
+  echo "Name=Files" >> /etc/xdg/autostart/nemo-autostart.desktop
+  echo "Exec=nemo-desktop" >> /etc/xdg/autostart/nemo-autostart.desktop
+  echo "OnlyShowIn=GNOME;Unity;" >> /etc/xdg/autostart/nemo-autostart.desktop
+  echo "X-Ubuntu-Gettext-Domain=nemo" >> /etc/xdg/autostart/nemo-autostart.desktop
 
-sudo apt install nemo -y
+  echo "Para que se aplique los cambios tiene que reiniciar el equipo [Y/n]: "
+  read opcion
 
-xdg-mime default nemo.desktop inode/directory application/x-gnome-saved-search
-
-gsettings set org.gnome.desktop.background show-desktop-icons false
-
-gsettings set org.nemo.desktop show-desktop-icons true
-
-echo "Ahora va a editar un fichero y tendra que introducir el siguiente \"codigo\" (dentro de este script) para seguir con la instalacion: "
-echo "ENTER para continuar: "
-read vacio
-$vacio > /dev/null
-
-$VIM > /dev/null || sudo vim /etc/xdg/autostart/nemo-autostart.desktop
-
-# Copia este contenido en el fichero:
-# [Desktop Entry]
-# Type=Application
-# Name=Files
-# Exec=nemo-desktop
-# OnlyShowIn=GNOME;Unity;
-# X-Ubuntu-Gettext-Domain=nemo
-
-# Reinicia el PC
-# reboot
+  if [ "$opcion" == "y" ] || [ "$opcion" == "Y" ] ; then
+    reboot
+  elif [ "$opcion" == "n" ] || [ "$opcion" == "N" ]; then
+    echo "No se ha reiniciado el equipo"
+    exit 0
+  else 
+    echo "Opcion no valida"
+    exit 1
+  fi
+else
+  echo "Para ejecutar el script tienes que ser root"
+fi
